+++ lib/rules/format-query-block.js
@@ -8,71 +8,19 @@
 // ------------------------------------------------------------------------------
 // Requirements
 // ------------------------------------------------------------------------------
-
 const prettier = require("prettier");
+const {
+  getPrettierDefaultOptions,
+  getPrettierRcOptions,
+  getMergedPrettierOptions
+} = require("../utils/getPrettierOptions");
+const getIndentOptions = require("../utils/getIndentOptions");
 
 // ------------------------------------------------------------------------------
 // Settings
 // ------------------------------------------------------------------------------
 const prettierParser = "graphql";
 const LINES = /[^\r\n\u2028\u2029]+(?:$|\r\n|[\r\n\u2028\u2029])/g;
-
-/**
- * Prettier's default options that Prettier settings
- * @return {Object} { optionsName: optionsDefaultValue }
- */
-const prettierDefaultOptions = prettier
-  .getSupportInfo(prettier.version)
-  .options.reduce((result, current) => {
-    result[current.name] = current.default;
-    return result;
-  }, {});
-
-/**
- * Get Prettier's options that user settings
- * @param {string} filePath File path that formatted with Prettier
- * @return {Object} { optionsName: optionsDefaultValue }
- */
-const getPrettierOptions = filePath => {
-  return prettier.resolveConfig.sync(filePath, {
-    editorconfig: true
-  });
-};
-
-/**
- * Get merged options that Prettier's default options and user settings options
- * @param {object} prettierDefaultOptions
- * @param {object} prettierOptions
- * @return {object} Merged options
- */
-const getMergedPrettierOptions = (prettierDefaultOptions, prettierOptions) => {
-  return { ...prettierDefaultOptions, ...prettierOptions };
-};
-
-/**
- * Get indent options
- * @param {object} mergedPrettierOptions
- * @return {object}
- */
-const getIndentOptions = mergedPrettierOptions => {
-  const defaultOptions = {
-    baseIndent: 0,
-    indentChar: " "
-  };
-
-  const { tabWidth, useTabs } = mergedPrettierOptions;
-
-  const result = {
-    baseIndent: tabWidth
-  };
-
-  if (useTabs) {
-    result.baseIndent = 1;
-    result.indentChar = "\t";
-  }
-
-  return { ...defaultOptions, ...result };
-};
 
 module.exports = {
   meta: {
@@ -91,8 +39,8 @@ module.exports = {
     const filePath = context.getFilename();
 
     const mergedPrettierOptions = getMergedPrettierOptions(
-      prettierDefaultOptions,
-      getPrettierOptions(filePath)
+      getPrettierDefaultOptions,
+      getPrettierRcOptions(filePath)
     );
 
     const { baseIndent, indentChar } = getIndentOptions(mergedPrettierOptions);
@@ -129,10 +77,17 @@ module.exports = {
               )
               .trim();
 
-            let lines = prettierCode.match(LINES);
+            let lines;
+            lines = prettierCode.match(LINES);
             lines = ["\n", ...lines, "\n"];
 
-            const formattedCode = lines.map(line => indent + line).join("");
+            const formattedCode = lines
+              .map(line => {
+                if (line === "\n") return line;
+                return indent + line;
+              })
+              .join("");
+            // const formattedCode = lines.map(line => indent + line).join("");
 
             if (formattedCode !== code) {
               context.report({
